feat(service) : Store Service 유저 검증 기능 추가.

+++ src/main/java/com/example/sparta/domain/store/service/StoreService.java
@@ -37,22 +37,36 @@ public class StoreService {
         return responseDtoList;
     }
 
-    public Long editStore(Long id, StoreRequestDto storeRequestDto) {
+    public Long editStore(Long id, StoreRequestDto storeRequestDto,User user) {
         Store store = storeRepository.findById(id).orElseThrow( ()->new NoSuchElementException("해당 가게를 찾을수 없어요."));
-        store.update(storeRequestDto);
+        if(checkingUserAccessPermission(user.getUserId(),store.getOwner().getUserId())) {
+            try {
+                store.update(storeRequestDto);
+            }
+            catch (Exception e) {
+                throw new IllegalArgumentException("수정 하는데 오류 발생");
+            }
+        }
         return id;
     }
 
-    public Long deleteStore(Long id){
+    public Long deleteStore(Long id,User user){
         Store store = storeRepository.findById(id).orElseThrow( ()->new NoSuchElementException("해당 가게를 찾을수 없어요."));
         try {
-            storeRepository.delete(store);
+            if(checkingUserAccessPermission(user.getUserId(),store.getOwner().getUserId())) {
+                storeRepository.delete(store);
+            }
         }catch (IllegalArgumentException e){
             throw new IllegalArgumentException("삭제 하는데 오류 발생");
         }
         return id;
     }
 
+    public StoreResponseDto getStoreById(Long id) {
+        Store store = storeRepository.findById(id).orElseThrow(()-> new NoSuchElementException("해당 가게를 찾을수 없어요."));
+        return new StoreResponseDto(store);
+    }
+    ///    EXTRA
     public List<StoreResponseDto> getAllStoreByName(String name) {
         List<Store> stores = storeRepository.findAllByNameContains(name);
         List<StoreResponseDto> responseDtoList = new ArrayList<>();
@@ -62,8 +76,20 @@ public class StoreService {
         return responseDtoList;
     }
 
-    public StoreResponseDto getStoreById(Long id) {
-        Store store = storeRepository.findById(id).orElseThrow(()-> new NoSuchElementException("해당 가게를 찾을수 없어요."));
-        return new StoreResponseDto(store);
+    /*
+    *  Finding the User.
+    *
+    */
+
+//    private User findUserById(long id){
+//        return userRepository.findById(id).orElseThrow(()-> new NoSuchElementException("해당 유저를 찾을수 없어요"));
+//    }
+    private boolean checkingUserAccessPermission(long userId,long storeId ){
+        if(userId==storeId){
+            return true;
+        }
+        else {
+            throw new IllegalArgumentException("이 게시글 에 수정 권한이 없습니다.");
+        }
     }
 }
